capturing user test in web env is somewhat challenging
Currently, we write to a new file that logs when user switch between chrome and code.

The hope is that they save a file and then test their web app on chrome, but currently it is hard to determine whether they really visit the localhost to test or if they are looking at other info to update their code. We currently wait for 4 secs before committing.

The log output using tee command is, on the other hand, synced correctly.